+++ CHANGELOG.md
@@ -1,13 +1,23 @@
+## 5.0.8
+
+- Updated native Mapp Intelligence SDK to version 5.1.11
+- Added missing methods to disable auto tracking, activity tracking and fragment tracking
+  - disableAutoTracking(bool); disableActivityTracking(bool); disableFragmentTracking(bool);
+- Fixed bug - setSendAppVersionInEveryRequest not resolved when called with await;
+  - issue: https://github.com/mapp-digital/Mapp-Intelligence-Flutter-Tracking/issues/8
+- Updated kotlin version: 2.0.20
+
 ## 5.0.7
+
 Fixed bug on Android - duplicate records are send for use case when application starts with a optOut=true and some tracking records exists in the local database.
 
 Fix fns feture iOS specific.
 
 **Bug Fixes**
-* App first opens were sometimes not tracked correctly.
-* Visits were not tracked as expected
-* If gallery permission dialog in iOS was used the SDK did not send the first track request.
 
+- App first opens were sometimes not tracked correctly.
+- Visits were not tracked as expected
+- If gallery permission dialog in iOS was used the SDK did not send the first track request.
 
 ## 5.0.6